refactor(dscache): Move hook to a subpackage of event
