+++ lib/lib.go
@@ -23,9 +23,9 @@ import (
 	"github.com/qri-io/qri/config"
 	"github.com/qri-io/qri/config/migrate"
 	"github.com/qri-io/qri/dscache"
-	"github.com/qri-io/qri/dsref/hook"
 	qrierr "github.com/qri-io/qri/errors"
 	"github.com/qri-io/qri/event"
+	"github.com/qri-io/qri/event/hook"
 	"github.com/qri-io/qri/fsi"
 	"github.com/qri-io/qri/fsi/hiddenfile"
 	"github.com/qri-io/qri/logbook"
